requirements.txt
pyrogram
tgcrypto
countryinfo
python-dotenv
